update and delete tasks

+++ src/services/task.service.js
@@ -13,7 +13,7 @@ async function create(taskObj) {
         const { valid, message } = tagsUtils.validateTags(taskObj.tags);
         if (!valid) {
             console.log(`Error validating tags when creating task: ${message}`);
-            return { data: null, error: {message: message} };
+            return { data: null, error: { message: message } };
         }
         const processedTags = tagsUtils.processTags(taskObj.tags);
         taskObj.tags = processedTags;
@@ -53,23 +53,26 @@ async function update(taskId, taskObj) {
         const { valid, message } = tagsUtils.validateTags(taskObj.etiquetas);
         if (!valid) {
             console.log(`Error validating tags when updating task with ID ${taskId}: ${message}`);
-            return { data: null, error: {message: message} };
+            return { data: null, error: { message: message } };
         }
         const processedTags = tagsUtils.processTags(taskObj.etiquetas);
         taskObj.etiquetas = processedTags;
     }
-    
-    let returnData = {message: "", data: {}};
+
+    taskObj.valorGemas = currenciesAndPoints.calculateGemPrice(taskObj.prioridad, taskObj.tiempo);
+    taskObj.puntosExperiencia = currenciesAndPoints.calculateExperiencePoints(taskObj.prioridad, taskObj.tiempo);
+
+    let returnData = { message: "", data: {} };
     const { data, error } = await supabase
         .from('Tareas')
         .update(taskObj)
         .eq('Tarea_ID', taskId);
-        if (error) {
-            console.log("Error updating task on supabase: ", error);
-        } else {
-            returnData.message = `Task with id ${taskId} updated with: ${JSON.stringify(taskObj, null, 2)}`;
-            returnData.data.taskId = taskId;
-        }
+    if (error) {
+        console.log("Error updating task on supabase: ", error);
+    } else {
+        returnData.message = `Task with id ${taskId} updated with: ${JSON.stringify(taskObj, null, 2)}`;
+        returnData.data.taskId = taskId;
+    }
     return { data: returnData, error };
 }
 
@@ -78,7 +81,7 @@ async function getById(taskId, columns = '*') {
         .from('Tareas')
         .select(columns)
         .eq('Tarea_ID', taskId);
-    
+
     error ? console.log(error) : console.log(`Task with ID ${taskId} found: ${JSON.stringify(data, null, 2)}`);
     return { data, error };
 }
@@ -93,7 +96,7 @@ async function formatTasks(data) {
     };
 
     const usersIds = data.map(task => task.Usuario_ID).filter(userId => userId !== null);
-    
+
     const { data: usersData, error: usersError } = await userService.getByIds(usersIds, 'Usuario_ID, nombre, apellido');
 
     if (usersError) {
@@ -116,8 +119,8 @@ async function formatTasks(data) {
             description: task.descripcion || "Sin descripción",
             projectName: `${task.Proyectos.nombre}`,
             tags: tags,
-            endDate: task.fechaFin 
-                ? new Date(task.fechaFin).toLocaleDateString('es-ES', { day: 'numeric', month: 'short', year: 'numeric' }) 
+            endDate: task.fechaFin
+                ? new Date(task.fechaFin).toLocaleDateString('es-ES', { day: 'numeric', month: 'short', year: 'numeric' })
                 : "Sin fecha"
         };
 
@@ -156,7 +159,7 @@ async function getByProjectId(projectId, columns = '*, Proyectos(nombre)', forma
 
 async function getTagsByProjectId(projectId) {
     const { data, error } = await getByProjectId(projectId, 'etiquetas');
-    
+
     if (error) {
         console.log(error);
         return { data: null, error };
@@ -174,7 +177,7 @@ async function getTagsByProjectId(projectId) {
     }
 }
 
-async function get(){
+async function get() {
     const { data, error } = await supabase
         .from('Tareas')
         .select()
@@ -182,24 +185,48 @@ async function get(){
     return { data, error };
 }
 
-async function deleteById(taskId) {
-    let returnData = {message: "", data: {}};
-    const { data, error } = await supabase
-        .from('Tareas')
-        .delete()
-        .eq('Tarea_ID', taskId);
+async function deleteById(taskId, userId, projectId) {
+    const LIDER_ROLE_NAME = "Lider"; // Nombre del rol líder
 
-        if (error) {
-            console.log("Error deleting task on supabase: ", error);
-        } else {
-            returnData.message = `Task with id ${taskId} deleted`;
-            returnData.data.taskId = taskId;
-        }
-    return { data: returnData, error };
+    // Verificar si el usuario tiene el rol de líder en el proyecto
+    const { data: hasRole, error: roleError } = await projectMemberService.checkMemberRole(userId, projectId, LIDER_ROLE_NAME);
+
+    if (roleError) {
+        console.error("Error al verificar el rol del usuario:", roleError);
+        return { data: null, error: roleError };
+    }
+
+    if (!hasRole) {
+        return {
+            data: null,
+            error: {
+                message: "Permiso denegado: solo los usuarios con el rol de líder pueden eliminar tareas.",
+                status: 403,
+            },
+        };
+    }
+
+    // Proceder con la eliminación si el usuario es líder
+    let returnData = { message: "", data: {} };
+    const { data, error } = await supabase
+        .from("Tareas")
+        .delete()
+        .eq("Tarea_ID", taskId);
+
+    if (error) {
+        console.error("Error al eliminar la tarea en Supabase:", error);
+        return { data: null, error };
+    }
+
+    returnData.message = `Tarea con ID ${taskId} eliminada correctamente.`;
+    returnData.data.taskId = taskId;
+
+    return { data: returnData, error: null };
 }
 
+
 async function updateTaskStatus(taskId, projectId, newStatusId, userId) {
-    let returnData = {message: "", data: {}};
+    let returnData = { message: "", data: {} };
     let task = null;
 
     const taskColumns = 'Tarea_ID, Estado_Tarea_ID, Usuario_ID, fueReclamada, puntosExperiencia, valorGemas';
@@ -214,11 +241,11 @@ async function updateTaskStatus(taskId, projectId, newStatusId, userId) {
     if (newStatusId === 4) {
         // Check if task has a user associated
         if (!task.Usuario_ID) {
-            return { data: null, error: {message: "Task has no user associated", status: 400} };
+            return { data: null, error: { message: "Task has no user associated", status: 400 } };
         }
         // Check if user has permission to approve tasks
         const { data: userData, error: userError } = await projectMemberService
-          .getByUserProject(userId, projectId, 'Usuario_ID, Proyecto_ID, Rol_ID, Roles (nombre)');
+            .getByUserProject(userId, projectId, 'Usuario_ID, Proyecto_ID, Rol_ID, Roles (nombre)');
         console.log(userData);
 
         if (userError) {
@@ -226,7 +253,7 @@ async function updateTaskStatus(taskId, projectId, newStatusId, userId) {
             return { data: null, error: userError };
         }
         if (userData[0].Roles.nombre !== 'Lider' && userData[0].Rol_ID !== 2) {
-            return { data: null, error: {message: "User does not have permission to approve tasks", status: 400} };
+            return { data: null, error: { message: "User does not have permission to approve tasks", status: 400 } };
         }
 
         // Call procedure to update task status
@@ -249,7 +276,7 @@ async function updateTaskStatus(taskId, projectId, newStatusId, userId) {
 
     }
     else if (newStatusId !== task.Estado_Tarea_ID) {
-        const { data, error: updateError } = await update(taskId, {Estado_Tarea_ID: newStatusId});
+        const { data, error: updateError } = await update(taskId, { Estado_Tarea_ID: newStatusId });
 
         if (updateError) {
             console.log("Error updating task status on supabase: ", updateError);
@@ -261,10 +288,10 @@ async function updateTaskStatus(taskId, projectId, newStatusId, userId) {
         }
     }
     else {
-        return { data: null, error: {message: "Task already has the new status", status: 400} };
+        return { data: null, error: { message: "Task already has the new status", status: 400 } };
     }
 
-    return { data: returnData, error: null};
+    return { data: returnData, error: null };
 }
 
 async function approvedTasksByProject(projectId, count = true) {
@@ -273,18 +300,103 @@ async function approvedTasksByProject(projectId, count = true) {
         .select()
         .eq('Proyecto_ID', projectId)
         .eq('Estado_Tarea_ID', 4);
-    
+
     if (error) {
         console.log(error);
         return { data: null, error };
     }
-    
+
     if (data && count) {
         return { data: { count: data.length }, error };
     }
 
     return { data, error };
 }
+
+async function updateByRole(taskObj, userId) {
+    const { Proyecto_ID, Task_ID, ...updates } = taskObj;
+
+    // Verificar si el usuario pertenece al proyecto y obtener su Rol_ID
+    const { data: memberData, error: memberError } = await projectMemberService.getByUserProject(
+        userId,
+        Proyecto_ID,
+        "Rol_ID"
+    );
+
+    if (memberError) {
+        console.error("Error al verificar el rol del usuario:", memberError);
+        return { data: null, error: memberError };
+    }
+
+    if (!memberData || memberData.length === 0) {
+        return {
+            data: null,
+            error: { message: "El usuario no es miembro del proyecto", status: 403 },
+        };
+    }
+
+    const userRoleId = memberData[0].Rol_ID;
+
+    // Definir los campos permitidos según el rol
+    let allowedFields = [];
+    if (userRoleId === 2) {
+        // Líder
+        allowedFields = [
+            "nombre",
+            "descripcion",
+            "prioridad",
+            "tiempo",
+            "etiquetas",
+            "gastos",
+            "presupuesto",
+            "Usuario_ID",
+            "fechaInicio",
+            "fechaFin",
+            "esCritica",
+
+        ];
+    } else if (userRoleId === 1) {
+        // Miembro
+        allowedFields = ["nombre", "descripcion", "etiquetas"];
+    } else {
+        return {
+            data: null,
+            error: { message: "Rol de usuario sin permisos para actualizar tareas", status: 403 },
+        };
+    }
+
+    // Filtrar los campos del objeto tarea según los permitidos
+    const filteredUpdates = Object.keys(updates)
+        .filter((key) => allowedFields.includes(key))
+        .reduce((obj, key) => {
+            obj[key] = updates[key];
+            return obj;
+        }, {});
+
+    if (Object.keys(filteredUpdates).length === 0) {
+        return {
+            data: null,
+            error: { message: "No hay campos válidos para actualizar", status: 400 },
+        };
+    }
+
+    // Usar el método `update` del servicio `taskService`
+    const { data, error } = await update(Task_ID, filteredUpdates);
+
+    if (error) {
+        console.error("Error al actualizar la tarea:", error);
+        return { data: null, error };
+    }
+
+    return {
+        data: {
+            message: `Tarea actualizada exitosamente con los campos: ${Object.keys(filteredUpdates).join(", ")}`,
+            Task_ID,
+        },
+        error: null,
+    };
+}
+
 
 export default {
     create,
@@ -296,4 +408,5 @@ export default {
     deleteById,
     updateTaskStatus,
     approvedTasksByProject,
+    updateByRole,
 };
